requirements.txt
selenium==4.9.*
beautifulsoup4==4.12.2
python-dotenv==1.0.*
requests==2.31.*
pyyaml==6.0.*
pyvirtualdisplay==3.0
pydantic==2.4.*
pyautogui==0.9.*